requirements.txt
pytest
allure-pytest
selene>=2.0.0rc9
Appium-Python-Client
requests
python-dotenv
pydantic